Exit codes(except redir&pipes)

+++ ctrlcz.c
@@ -13,9 +13,13 @@ void CTRLZ0(int sig)
 // to stop running foreground
 void CTRLZ(int sig)
 {
-    kill(fore_pid,SIGTSTP);
-    bg_jobs[bg_cnt]=fore_pid;
-    bg_cnt++;
-    printf("\n\033[1;31mStopped: [%lld]\033[0m\n",fore_pid);
+    if (fore_pid>0)
+    {
+        kill(fore_pid,SIGTSTP);
+        bg_jobs[bg_cnt]=fore_pid;
+        bg_cnt++;
+        printf("\n\033[1;31mStopped: [%lld]\033[0m\n",fore_pid);
+        fore_pid=-1;
+    }
     return;
 }
